# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: InterHive API
  description: InterHive - Industry Readiness & Talent Connect Platform API
  version: 1.0.0
  contact:
    name: InterHive Team
    email: support@interhive.in
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.interhive.in/api/v1
    description: Production Server
  - url: https://staging-api.interhive.in/api/v1
    description: Staging Server
  - url: http://localhost:3000/api/v1
    description: Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InternProfile:
      type: object
      required:
        - personalInfo
        - contact
        - academicInfo
        - preferences
      properties:
        personalInfo:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            dateOfBirth:
              type: string
              format: date
            gender:
              type: string
              enum: [male, female, other]
            nationality:
              type: string
            profilePhoto:
              type: string
        contact:
          type: object
          properties:
            email:
              type: string
              format: email
            phone:
              type: string
            alternatePhone:
              type: string
            address:
              type: object
              properties:
                street:
                  type: string
                city:
                  type: string
                state:
                  type: string
                country:
                  type: string
                zipCode:
                  type: string
            socialMedia:
              type: object
              properties:
                linkedin:
                  type: string
                github:
                  type: string
                portfolio:
                  type: string
        academicInfo:
          type: object
          properties:
            currentEducation:
              type: object
              properties:
                institution:
                  type: string
                degree:
                  type: string
                field:
                  type: string
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                isCurrent:
                  type: boolean
                grade:
                  type: string
            cgpa:
              type: number
              minimum: 0
              maximum: 10
            graduationYear:
              type: integer
        preferences:
          type: object
          properties:
            preferredDomains:
              type: array
              items:
                type: string
            preferredLocation:
              type: array
              items:
                type: string
            preferredWorkType:
              type: array
              items:
                type: string
                enum: [remote, hybrid, onsite]
            expectedStipend:
              type: object
              properties:
                min:
                  type: number
                max:
                  type: number
    ReadinessScore:
      type: object
      properties:
        overall:
          type: integer
          minimum: 0
          maximum: 100
        breakdown:
          type: object
          properties:
            technicalSkills:
              type: integer
            projects:
              type: integer
            communication:
              type: integer
            problemSolving:
              type: integer
            industryWorkflow:
              type: integer
            teamCollaboration:
              type: integer
            leadership:
              type: integer
            adaptability:
              type: integer
        lastUpdated:
          type: string
          format: date-time

    # Company
...
paths:
...
  /interns/profile:
    get:
      tags: [Interns]
      summary: Get intern profile
      security:
        - bearerAuth: []
      responses:
        200:
          description: Profile retrieved
        404:
          description: Profile not found

    post:
      tags: [Interns]
      summary: Create intern profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternProfile'
      responses:
        201:
          description: Profile created

    put:
      tags: [Interns]
      summary: Update intern profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternProfile'
      responses:
        200:
          description: Profile updated
  /interns/readiness:
    get:
      tags: [Interns]
      summary: Get readiness score
      security:
        - bearerAuth: []
      responses:
        200:
          description: Readiness score retrieved
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ReadinessScore'
  /interns/applications:
    post:
      tags: [Interns]
      summary: Apply for program
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                programId:
                  type: string
                coverLetter:
                  type: string
      responses:
        201:
          description: Application submitted
  /interns/opportunities:
    get:
      tags: [Interns]
      summary: Get matched opportunities
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Opportunities retrieved

  # Companies
...
  /matching/find-matches:
    post:
      tags: [Matching]
      summary: Find matches
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                minScore:
                  type: integer
                  minimum: 0
                  maximum: 100
                limit:
                  type: integer
      responses:
        200:
          description: Matches found
  /matching/my-matches:
    get:
      tags: [Matching]
      summary: Get my matches
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, interview_scheduled, interview_completed, offer_made, offer_accepted, offer_rejected, hired, expired]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Matches retrieved
  /matching/{id}/accept:
    post:
      tags: [Matching]
      summary: Accept match
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Match accepted
  /matching/{id}/reject:
    post:
      tags: [Matching]
      summary: Reject match
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Match rejected
  /matching/{id}/schedule-interview:
    post:
      tags: [Matching]
      summary: Schedule interview
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - interviewDate
                - interviewType
              properties:
                interviewDate:
                  type: string
                  format: date-time
                interviewType:
                  type: string
                meetingLink:
                  type: string
      responses:
        200:
          description: Interview scheduled

  # Attendance
...